webkit-efl/LayoutTests/platform/efl-tizen/webaudio/test-basic-expected.txt

CONSOLE MESSAGE: line 42: ReferenceError: Can't find variable: webkitAudioContext
Tests basic audio test infrastructure: specifically layoutTestController.setAudioData()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
